refactoring of MLFlowBasedModel and WrapperModel
